requirements.txt
python-telegram-bot==21.10
apscheduler==3.10.4
python-dateutil==2.8.2